refactor(ui): extract comparison legend component and improve debug section visibility
